requirements.txt
# ---- requirements.txt ----
gradio
transformers
Pillow
requests
torch